Renamed single frame value to per frame value, moved it to its own file, and added comments for clarity.
